+++ examples/example_actor_per_row.rs
@@ -89,7 +89,7 @@ impl actix::Handler<woab::Signal> for AddendActor {
     fn handle(&mut self, msg: woab::Signal, ctx: &mut Self::Context) -> Self::Result {
         Ok(match msg.name() {
             "addend_changed" => {
-                let buffer: gtk::TextBuffer = msg.param(0)?;
+                let woab::params!(buffer: gtk::TextBuffer) = msg.params()?;
                 let new_number = buffer
                     .get_text(&buffer.get_start_iter(), &buffer.get_end_iter(), true)
                     .and_then(|text| text.parse().ok());