+++ macros/src/lib.rs
@@ -2,6 +2,7 @@ mod factories_derive;
 mod removable_derive;
 mod util;
 mod widgets_from_builder_derive;
+mod param_extraction;
 
 #[proc_macro_derive(WidgetsFromBuilder, attributes(widget))]
 pub fn derive_widgets_from_builder(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
@@ -25,6 +26,15 @@ pub fn derive_factories(input: proc_macro::TokenStream) -> proc_macro::TokenStre
 pub fn derive_removable(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let input = syn::parse_macro_input!(input as syn::DeriveInput);
     match removable_derive::impl_removable_derive(&input) {
+        Ok(output) => output.into(),
+        Err(error) => error.to_compile_error().into(),
+    }
+}
+
+#[proc_macro]
+pub fn params(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    let input = syn::parse_macro_input!(input as param_extraction::Input);
+    match input.impl_param_extraction() {
         Ok(output) => output.into(),
         Err(error) => error.to_compile_error().into(),
     }