+++ src/signal.rs
@@ -9,7 +9,7 @@ impl<T> actix::Message for Signal<T> {
     type Result = SignalResult;
 }
 
-struct SignalData<T> {
+pub struct SignalData<T> {
     name: Rc<String>,
     parameters: Vec<glib::Value>,
     tag: T,
@@ -28,6 +28,34 @@ impl<T: Clone> Signal<T> {
     }
 }
 
+#[doc(hidden)]
+impl<T> SignalData<T> {
+    fn raw_param<'a>(&'a self, index: usize) -> Result<&'a glib::Value, crate::Error> {
+        self
+            .parameters
+            .get(index)
+            .ok_or_else(|| crate::Error::SignalParameterIndexOutOfBound {
+                signal: self.name.as_str().to_owned(),
+                index,
+                num_parameters: self.parameters.len(),
+            })
+    }
+
+    fn param<'a, P: glib::value::FromValueOptional<'a>>(&'a self, index: usize) -> Result<P, crate::Error> {
+        let value = self.raw_param(index)?;
+        if let Ok(Some(value)) = value.get() {
+            Ok(value)
+        } else {
+            Err(crate::Error::IncorrectSignalParameter {
+                signal: self.name.as_str().to_owned(),
+                index,
+                expected_type: <P as glib::types::StaticType>::static_type(),
+                actual_type: value.type_(),
+            })
+        }
+    }
+}
+
 impl<T> Signal<T> {
     pub fn new(name: Rc<String>, parameters: Vec<glib::Value>, tag: T) -> Self {
         Signal(SendWrapper::new(SignalData { name, parameters, tag }))
@@ -42,25 +70,7 @@ impl<T> Signal<T> {
     }
 
     pub fn param<'a, P: glib::value::FromValueOptional<'a>>(&'a self, index: usize) -> Result<P, crate::Error> {
-        let value = &self
-            .0
-            .parameters
-            .get(index)
-            .ok_or_else(|| crate::Error::SignalParameterIndexOutOfBound {
-                signal: self.name().to_owned(),
-                index,
-                num_parameters: self.0.parameters.len(),
-            })?;
-        if let Ok(Some(value)) = value.get() {
-            Ok(value)
-        } else {
-            Err(crate::Error::IncorrectSignalParameter {
-                signal: self.name().to_owned(),
-                index,
-                expected_type: <P as glib::types::StaticType>::static_type(),
-                actual_type: value.type_(),
-            })
-        }
+        self.0.param(index)
     }
 
     pub fn action_param<P: glib::variant::FromVariant>(&self) -> Result<P, crate::Error> {
@@ -74,5 +84,45 @@ impl<T> Signal<T> {
 
     pub fn cant_handle(&self) -> SignalResult {
         Err(crate::Error::NoSuchSignalError("Actor", (*self.0.name).to_owned()))
+    }
+
+    pub fn params<'a, R: SignalParamReceiver<'a>>(&'a self) -> Result<R, crate::Error> {
+        R::fill_from_index(&*self.0, 0)
+    }
+}
+
+pub trait SignalParamReceiver<'a>: Sized {
+    fn fill_from_index<D>(signal: &'a SignalData<D>, from_index: usize) -> Result<Self, crate::Error>;
+}
+
+impl SignalParamReceiver<'_> for () {
+    fn fill_from_index<D>(signal: &SignalData<D>, from_index: usize) -> Result<Self, crate::Error> {
+        if from_index < signal.parameters.len() {
+            return Err(crate::Error::NotAllParametersExtracted {
+                signal: signal.name.as_str().to_owned(),
+                num_parameters: signal.parameters.len(),
+                num_extracted: from_index,
+            })
+        }
+        Ok(())
+    }
+}
+
+impl<'a, T, R> SignalParamReceiver<'a> for (T, core::marker::PhantomData<T>, R)
+where
+    T: glib::value::FromValueOptional<'a>,
+    R: SignalParamReceiver<'a>,
+{
+    fn fill_from_index<D>(signal: &'a SignalData<D>, from_index: usize) -> Result<Self, crate::Error> {
+        Ok((signal.param(from_index)?, core::marker::PhantomData, R::fill_from_index(signal, from_index + 1)?))
+    }
+}
+
+impl<'a, R> SignalParamReceiver<'a> for ((&'a glib::Value,), R)
+where
+    R: SignalParamReceiver<'a>,
+{
+    fn fill_from_index<D>(signal: &'a SignalData<D>, from_index: usize) -> Result<Self, crate::Error> {
+        Ok(((signal.raw_param(from_index)?,), R::fill_from_index(signal, from_index + 1)?))
     }
 }